Switching from underscore to lodash
Lodash is a drop-in replacement for underscore that is faster and
more actively maintained. Since it is drop-in compatible, switching to it
is trivial. The `require` statements are just changed from lines like
`_ = require 'underscore'` to `_ = require 'lodash-node'`

Tested plan: Ran `cake test` and manually tested a few things

See:
http://joefleming.net/posts/use-lodash-instead-of-underscore/
http://stackoverflow.com/questions/13789618/differences-between-lodash-and-underscore