Switch the doc plugin to use the nesh-doc npm package
This should make it easier to update that separately, while
still allowing the nesh project to adopt new versions at its
own pace.

Also, keeps the dependencies of that separated out from
the ones nesh depends on (was able to remove intdoc and
lodash from nesh's package.json with this change).